Move CLientState, ConsoleClient, TorrentClient to torrent.client package
