+++ src/main/java/net/yeputons/spbau/spring2016/torrent/Main.java
@@ -1,5 +1,7 @@
 package net.yeputons.spbau.spring2016.torrent;
 
+import net.yeputons.spbau.spring2016.torrent.client.ConsoleClient;
+
 import java.io.IOException;
 import java.nio.file.Paths;
 import java.util.Arrays;
